refactor(parser): split object literal parsing recovery
