+++ crates/tsz-parser/src/parser/state_expressions_literals.rs
@@ -1602,163 +1602,6 @@ impl ParserState {
         }
     }
 
-    /// Whether the current token can start an object-literal element, mirroring
-    /// tsc's `isListElement(ParsingContext.ObjectLiteralMembers)`:
-    /// `[` (computed), `*` (generator), `...` (spread), `.` (kept so the object
-    /// is not closed mid-completion), or any literal property name
-    /// (identifier/keyword/string/numeric/bigint/`#private`). Tokens outside
-    /// this set are not members and are skipped by the member loop's recovery.
-    ///
-    /// Template literals are also routed here so that the dedicated
-    /// template-literal-property recovery inside `parse_property_assignment`
-    /// (which closes the object and recovers the template as a tagged-template
-    /// tail) keeps running instead of being skipped by the generic recovery.
-    pub(crate) const fn is_object_literal_element_start(&self) -> bool {
-        matches!(
-            self.token(),
-            SyntaxKind::OpenBracketToken
-                | SyntaxKind::AsteriskToken
-                | SyntaxKind::DotDotDotToken
-                | SyntaxKind::DotToken
-                | SyntaxKind::NoSubstitutionTemplateLiteral
-                | SyntaxKind::TemplateHead
-        ) || self.is_property_name()
-    }
-
-    /// Parse object literal
-    pub(crate) fn parse_object_literal(&mut self) -> NodeIndex {
-        let start_pos = self.token_pos();
-        self.parse_expected(SyntaxKind::OpenBraceToken);
-
-        let mut properties = Vec::new();
-        while !self.is_token(SyntaxKind::CloseBraceToken) {
-            // Mirror tsc's `parseDelimitedList(ObjectLiteralMembers)`:
-            // before parsing a member, a token that cannot start an object
-            // literal element (and is not `}` / `;` / EOF, which are handled
-            // by the terminator/separator logic below) is reported via
-            // `parsingContextErrors` (TS1136 "Property assignment expected.")
-            // and skipped — it never produces a member. Without this, a stray
-            // operator/punctuation taken as a property name (e.g. the `:` left
-            // after recovering `x()?: 1`) would be folded into a bogus
-            // empty-named member instead of being skipped like tsc.
-            if !self.is_object_literal_element_start()
-                && !self.is_token(SyntaxKind::SemicolonToken)
-                && !self.is_token(SyntaxKind::EndOfFileToken)
-            {
-                use tsz_common::diagnostics::diagnostic_codes;
-                self.parse_error_at_current_token(
-                    "Property assignment expected.",
-                    diagnostic_codes::PROPERTY_ASSIGNMENT_EXPECTED,
-                );
-                self.next_token();
-                self.suppress_object_literal_comma_once = true;
-                continue;
-            }
-            let prop = self.parse_property_assignment();
-            if prop.is_some() {
-                properties.push(prop);
-            }
-            if self.abort_object_literal_recovery_once {
-                self.abort_object_literal_recovery_once = false;
-                break;
-            }
-
-            // Try to parse comma separator
-            if !self.parse_optional(SyntaxKind::CommaToken) {
-                if self.suppress_object_literal_comma_once && self.is_property_start() {
-                    self.suppress_object_literal_comma_once = false;
-                    continue;
-                }
-                self.suppress_object_literal_comma_once = false;
-
-                if self.is_token(SyntaxKind::SemicolonToken) {
-                    // Semicolons in object literals: look ahead to decide whether to
-                    // treat as a mistyped comma (continue) or abort the list (break).
-                    // tsc's parseDelimitedList aborts when the token after `;` is in
-                    // some other parsing context (e.g., EOF, statement keyword without
-                    // subsequent property-like content). We look ahead past `;` to
-                    // decide: if the next token looks like it could continue the object
-                    // literal (property start, or `}` to close it), treat `;` as a
-                    // mistyped comma. Otherwise, abort the list so the outer parser
-                    // can handle the rest as statements.
-                    let snapshot = self.scanner.save_state();
-                    let saved_token = self.current_token;
-                    self.next_token(); // look past `;`
-                    let should_continue =
-                        self.is_property_start() || self.is_token(SyntaxKind::CloseBraceToken);
-                    let follows_eof = self.is_token(SyntaxKind::EndOfFileToken);
-                    self.scanner.restore_state(snapshot);
-                    self.current_token = saved_token;
-
-                    if should_continue {
-                        // Treat `;` as mistyped `,` — emit error and continue
-                        use tsz_common::diagnostics::diagnostic_codes;
-                        self.parse_error_at_current_token(
-                            "',' expected.",
-                            diagnostic_codes::EXPECTED,
-                        );
-                        self.next_token(); // skip `;`
-                    } else if follows_eof {
-                        // tsc emits ',' expected at the `;` position
-                        // (its delimited-list parser reports the expected comma
-                        // before it knows whether the list is also unclosed).
-                        let diag_count_before = self.parse_diagnostics.len();
-                        self.error_comma_expected();
-                        let comma_error_emitted = self.parse_diagnostics.len() > diag_count_before;
-                        self.next_token(); // skip `;`, now at EOF
-                        if comma_error_emitted {
-                            // The comma error was actually emitted at `;`.
-                            // In tsc, both the comma error and the subsequent
-                            // close-brace error land at the same position
-                            // (because tsc's abortParsingList doesn't consume
-                            // the token). Set last_error_pos to EOF so that
-                            // parse_expected(CloseBrace) sees the same position
-                            // and deduplicates.
-                            self.last_error_pos = self.token_pos();
-                        }
-                        break;
-                    } else {
-                        // `;` followed by non-property → abort the list
-                        // so the outer parser can handle the rest as statements.
-                        break;
-                    }
-                } else if self.is_property_start() && !self.is_token(SyntaxKind::CloseBraceToken) {
-                    // We have a property-like token but no comma - likely missing comma
-                    // Emit the comma error and continue parsing for better recovery
-                    // This handles cases like: {a: 1 b: 2} instead of {a: 1, b: 2}
-                    self.error_comma_expected();
-                } else if self.is_token(SyntaxKind::EndOfFileToken)
-                    || self.is_token(SyntaxKind::CloseBraceToken)
-                {
-                    break;
-                } else {
-                    // Bypass `error_comma_expected`'s 3-byte distance gate so the
-                    // recovery emits even when the separator sits 3 cols after a
-                    // prior emission (e.g. `{` after `:' expected.` at 3 cols out
-                    // in `{ class C4 {} }`). tsc's parseErrorAtPosition only
-                    // dedups exact same-position duplicates; `parse_error_at`'s
-                    // dedup mirrors that.
-                    use tsz_common::diagnostics::diagnostic_codes;
-                    self.parse_error_at_current_token("',' expected.", diagnostic_codes::EXPECTED);
-                    self.next_token();
-                }
-            }
-        }
-
-        let end_pos = self.token_end();
-        self.parse_expected(SyntaxKind::CloseBraceToken);
-
-        self.arena.add_literal_expr(
-            syntax_kind_ext::OBJECT_LITERAL_EXPRESSION,
-            start_pos,
-            end_pos,
-            LiteralExprData {
-                elements: self.make_node_list(properties),
-                multi_line: false,
-            },
-        )
-    }
-
     /// Parse property assignment, method, getter, setter, or spread element
     pub(crate) fn parse_property_assignment(&mut self) -> NodeIndex {
         let start_pos = self.token_pos();
@@ -2667,27 +2510,6 @@ impl ParserState {
                 parameters,
                 type_annotation,
                 body,
-            },
-        )
-    }
-
-    /// Create a distinct empty/missing expression node to use as the value of a
-    /// property assignment whose initializer is absent (`{ x: }`). It is a
-    /// zero-length empty identifier at `pos`, separate from the property name so
-    /// the emitter does not mistake the property for a shorthand (which keys on
-    /// `name == initializer`) and instead renders `name:` with an empty value,
-    /// matching tsc.
-    fn create_missing_property_value(&mut self, pos: u32) -> NodeIndex {
-        let atom = self.scanner.interner_mut().intern("");
-        self.arena.add_identifier(
-            SyntaxKind::Identifier as u16,
-            pos,
-            pos,
-            IdentifierData {
-                atom,
-                escaped_text: String::new(),
-                original_text: None,
-                type_arguments: None,
             },
         )
     }
